refactor: decompose the components
