+++ src/components/feed-orders/feed-orders.module.css
@@ -25,4 +25,9 @@
 	border-radius: 6px;			
 }
 
+.link {
+	text-decoration: none;
+	color: #F2F2F3;
+}
+
