+++ src/components/forms-profile/forms-profile.module.css
@@ -1,0 +1,5 @@
+.forms {
+	display: flex;
+	flex-direction: column;
+	gap: 24px;
+}
