+++ src/components/orders-item/orders-item.module.css
@@ -58,5 +58,16 @@
   z-index: 3;
 }
 
+.status_color_red {
+	color: #E52B1A;
+}
+
+.status_color_white {
+	color: #F2F2F3;
+}
+
+.status_color_green {
+	color: #00CCCC;
+}
