+++ src/pages/order/order.module.css
@@ -48,7 +48,7 @@
 }
 
 .status {
-	color: #00CCCC;
+	/* //color: #00CCCC; */
 	text-align: start;
 }
 
@@ -70,4 +70,16 @@
 	background: #8585AD;
 	border: 2px solid #2F2F37;
 	border-radius: 6px;			
+}
+
+.status_color_red {
+	color: #E52B1A;
+}
+
+.status_color_white {
+	color: #F2F2F3;
+}
+
+.status_color_green {
+	color: #00CCCC;
 }
